localdumbmodel/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
llama-cpp-python  # Adjust version as needed; install with pip install llama-cpp-python (may require build dependencies like cmake)
